HTBHF-2507 move states to top level of flow control directory
