+++ src/web/routes/application/flow-control/index.js
@@ -1,9 +1,11 @@
 const { getPreviousPath } = require('./get-previous-path')
 const testUtils = require('./test-utils')
+const states = require('./states')
 
 module.exports = {
   ...require('./state-machine'),
   ...require('./middleware'),
   getPreviousPath,
-  testUtils
+  testUtils,
+  states
 }
