+++ src/web/routes/application/flow-control/state-machine/index.js
@@ -1,9 +1,7 @@
 const { stateMachine } = require('./state-machine')
-const states = require('./states')
 const actions = require('./actions')
 
 module.exports = {
-  states,
   actions,
   stateMachine
 }
